spilt files
