+++ projects/bare_metal/dynamic_memory.h
@@ -1,0 +1,18 @@
+/******************************************************************************
+*
+* This file contains functions that help facilitate dyanmic memory allocation
+*
+*******************************************************************************/
+
+#ifndef _DYNAMIC_MEM_ALLOCATION_
+#define _DYNAMIC_MEM_ALLOCATION_
+
+#include "stdint.h"
+#include "stdlib.h"
+#include "linker_defines.h"
+
+void * _sbrk(uint32_t incr);
+
+
+
+#endif
